verify state roots after Turing transactions
